Add support for a default prompt on tinymce rich text fields.
If a default_prompt argument is in keyword arguments to
exe.engine.common richTextField an overlay will be generated
to prompt the user (e.g. type the feedback when the user selects
this answer).

Added default_prompt property to TextAreaField that is used to set
this behaviour.

The prompt will disappear as soon as the user clicks or starts
typing in the field.