Keep client handles for whole session, show user when disconnected
eXeClientHandle does not get deleted when the client disappers the
first time - e.g. if the user just how a short interruption in their
connectivity.  They are expired and removed when the session expires
instead.

User will be shown when he/she is disconnected, UI will be disabled
until connection with server is re-established.

Occassionally on first load selection of first node not cooperating,
add check function in OutlineXmlTreeStore.js