Fix Issue #18 style change not showing in preview
Style changes were not being shown because last-modified headers were
set according to the modified date; and the user can switch to a style
that was modified before the current style; making the browser think
it has the most up to date copy.

Add support to validation by ETag.  In case of etag match invalidate
checking according to last modified date.